Move all assets into the project
This means that we can access the site entirely locally (i.e. without an
internet connection) instead of having to rely on the network. In the
future we should move these things into some form of controlled system
like bower + CDN w/ local fallbacks, but for now this is a marked
improvement.